implement "includeLabels" node processor
