+++ packages/kbn-openapi-bundler/src/__test__/include_labels.test.ts
@@ -9,12 +9,14 @@
 import { bundleSpecs } from './bundle_specs';
 import { createOASDocument } from './create_oas_document';
 
-describe('OpenAPI Bundler - exclude labeled nodes', () => {
-  it('excludes ESS endpoints', async () => {
+describe('OpenAPI Bundler - include labeled operations', () => {
+  it('includes Serverless endpoints', async () => {
     const spec = createOASDocument({
       paths: {
         '/api/some_api': {
           get: {
+            // @ts-expect-error custom property is unexpected here
+            'x-labels': ['serverless'],
             responses: {
               '200': {
                 description: 'Successful response',
@@ -29,7 +31,81 @@ describe('OpenAPI Bundler - exclude labeled nodes', () => {
             },
           },
           post: {
+            'x-labels': ['ess'],
+            responses: {
+              '200': {
+                description: 'Successful response',
+                content: {
+                  'application/json': {
+                    schema: {
+                      type: 'string',
+                    },
+                  },
+                },
+              },
+            },
+          },
+        },
+        '/api/another_api': {
+          get: {
             // @ts-expect-error custom property is unexpected here
+            'x-labels': ['ess'],
+            responses: {
+              '200': {
+                description: 'Successful response',
+                content: {
+                  'application/json': {
+                    schema: {
+                      type: 'string',
+                    },
+                  },
+                },
+              },
+            },
+          },
+        },
+      },
+    });
+
+    const [bundledSpec] = Object.values(
+      await bundleSpecs(
+        {
+          1: spec,
+        },
+        {
+          includeLabels: ['serverless'],
+        }
+      )
+    );
+
+    expect(bundledSpec.paths).toEqual({
+      '/api/some_api': {
+        get: expect.objectContaining({}),
+      },
+    });
+  });
+
+  it('includes ESS endpoints', async () => {
+    const spec = createOASDocument({
+      paths: {
+        '/api/some_api': {
+          get: {
+            // @ts-expect-error custom property is unexpected here
+            'x-labels': ['serverless'],
+            responses: {
+              '200': {
+                description: 'Successful response',
+                content: {
+                  'application/json': {
+                    schema: {
+                      type: 'string',
+                    },
+                  },
+                },
+              },
+            },
+          },
+          post: {
             'x-labels': ['ess'],
             responses: {
               '200': {
@@ -72,13 +148,16 @@ describe('OpenAPI Bundler - exclude labeled nodes', () => {
           1: spec,
         },
         {
-          excludeXLabels: ['ess'],
+          includeLabels: ['ess'],
         }
       )
     );
 
     expect(bundledSpec.paths).toEqual({
       '/api/some_api': {
+        post: expect.objectContaining({}),
+      },
+      '/api/another_api': {
         get: expect.objectContaining({}),
       },
     });
