+++ packages/kbn-openapi-bundler/src/bundler/bundle_document.ts
@@ -7,24 +7,25 @@
  */
 
 import { isAbsolute } from 'path';
-import { RefResolver } from './ref_resolver';
-import { processDocument } from './process_document';
-import { BundleRefProcessor } from './document_processors/bundle_refs';
-import { createSkipNodeWithInternalPropProcessor } from './document_processors/skip_node_with_internal_prop';
-import { createModifyPartialProcessor } from './document_processors/modify_partial';
-import { createSkipInternalPathProcessor } from './document_processors/skip_internal_path';
-import { ResolvedDocument, ResolvedRef } from './types';
-import { createRemovePropsProcessor } from './document_processors/remove_props';
-import { createModifyRequiredProcessor } from './document_processors/modify_required';
+import { RefResolver } from './ref_resolver/ref_resolver';
+import { processDocument } from './process_document/process_document';
 import { X_CODEGEN_ENABLED, X_INLINE, X_INTERNAL, X_MODIFY } from './known_custom_props';
-import { RemoveUnusedComponentsProcessor } from './document_processors/remove_unused_components';
 import { isPlainObjectType } from '../utils/is_plain_object_type';
+import { ResolvedDocument } from './ref_resolver/resolved_document';
+import { ResolvedRef } from './ref_resolver/resolved_ref';
+import { createSkipNodeWithInternalPropProcessor } from './process_document/document_processors/skip_node_with_internal_prop';
+import { createSkipInternalPathProcessor } from './process_document/document_processors/skip_internal_path';
+import { createModifyPartialProcessor } from './process_document/document_processors/modify_partial';
+import { createModifyRequiredProcessor } from './process_document/document_processors/modify_required';
+import { createRemovePropsProcessor } from './process_document/document_processors/remove_props';
 import {
   createFlattenFoldedAllOfItemsProcessor,
   createMergeNonConflictingAllOfItemsProcessor,
   createUnfoldSingleAllOfItemProcessor,
-} from './document_processors/reduce_all_of_items';
-import { createExcludeXLabelsProcessor } from './document_processors/exclude_x_labels';
+} from './process_document/document_processors/reduce_all_of_items';
+import { createIncludeXLabelsProcessor } from './process_document/document_processors/include_x_labels';
+import { BundleRefProcessor } from './process_document/document_processors/bundle_refs';
+import { RemoveUnusedComponentsProcessor } from './process_document/document_processors/remove_unused_components';
 
 export class SkipException extends Error {
   constructor(public documentPath: string, message: string) {
@@ -37,7 +38,7 @@ export interface BundledDocument extends ResolvedDocument {
 }
 
 interface BundleDocumentOptions {
-  excludeXLabels?: string[];
+  includeXLabels?: string[];
 }
 
 /**
@@ -88,8 +89,8 @@ export async function bundleDocument(
     createUnfoldSingleAllOfItemProcessor(),
   ];
 
-  if (options?.excludeXLabels) {
-    defaultProcessors.push(createExcludeXLabelsProcessor(options?.excludeXLabels));
+  if (options?.includeXLabels) {
+    defaultProcessors.push(createIncludeXLabelsProcessor(options?.includeXLabels));
   }
 
   const bundleRefsProcessor = new BundleRefProcessor(X_INLINE);
