+++ packages/kbn-openapi-bundler/src/bundler/document.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+/**
+ * Document abstraction. We don't mind OpenAPI `3.0` and `3.1` differences.
+ */
+export type Document = Record<string, unknown>;
