+++ packages/kbn-openapi-bundler/src/bundler/process_document/document_processors/modify_partial.ts
@@ -6,10 +6,10 @@
  * Side Public License, v 1.
  */
 
-import { DocumentNodeProcessor } from '../types';
-import { hasProp } from '../../utils/has_prop';
+import { DocumentNodeProcessor } from './types/document_node_processor';
+import { hasProp } from '../../../utils/has_prop';
+import { X_MODIFY } from '../../known_custom_props';
 import { inlineRef } from './utils/inline_ref';
-import { X_MODIFY } from '../known_custom_props';
 
 /**
  * Creates a node processor to modify a node by removing `required` property when
