+++ packages/kbn-openapi-bundler/src/bundler/process_document/document_processors/modify_required.ts
@@ -7,11 +7,11 @@
  */
 
 import chalk from 'chalk';
-import { logger } from '../../logger';
-import { isPlainObjectType } from '../../utils/is_plain_object_type';
-import { DocumentNodeProcessor } from '../types';
-import { hasProp } from '../../utils/has_prop';
-import { X_MODIFY } from '../known_custom_props';
+import { logger } from '../../../logger';
+import { isPlainObjectType } from '../../../utils/is_plain_object_type';
+import { hasProp } from '../../../utils/has_prop';
+import { X_MODIFY } from '../../known_custom_props';
+import { DocumentNodeProcessor } from './types/document_node_processor';
 import { inlineRef } from './utils/inline_ref';
 
 /**
