+++ packages/kbn-openapi-bundler/src/bundler/process_document/document_processors/reduce_all_of_items/flatten_folded_all_of_items.ts
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-import { DocumentNodeProcessor } from '../../types';
+import { DocumentNodeProcessor } from '../types/document_node_processor';
 
 /**
  * Creates a node processor to flatten folded `allOf` items. Folded means `allOf` has items
