+++ packages/kbn-openapi-bundler/src/bundler/process_document/document_processors/reduce_all_of_items/index.ts
