+++ packages/kbn-openapi-bundler/src/bundler/process_document/document_processors/reduce_all_of_items/unfold_single_all_of_item.ts
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-import { DocumentNodeProcessor } from '../../types';
+import { DocumentNodeProcessor } from '../types/document_node_processor';
 
 /**
  * Created a node processor to remove/unfold `allOf` with only single item.
