+++ packages/kbn-openapi-bundler/src/bundler/process_document/document_processors/remove_props.ts
@@ -6,8 +6,8 @@
  * Side Public License, v 1.
  */
 
-import { isPlainObjectType } from '../../utils/is_plain_object_type';
-import { DocumentNodeProcessor } from '../types';
+import { isPlainObjectType } from '../../../utils/is_plain_object_type';
+import { DocumentNodeProcessor } from './types/document_node_processor';
 
 /**
  * Creates a node processor to remove specified by `propNames` properties.
