+++ packages/kbn-openapi-bundler/src/bundler/process_document/document_processors/skip_node_with_internal_prop.ts
@@ -1,0 +1,21 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+import { DocumentNodeProcessor } from './types/document_node_processor';
+
+/**
+ * Creates a node processor to skip nodes having provided `skipProperty` property
+ * and omit them from the result document.
+ */
+export function createSkipNodeWithInternalPropProcessor(
+  skipProperty: string
+): DocumentNodeProcessor {
+  return {
+    shouldRemove: (node) => skipProperty in node,
+  };
+}
