+++ packages/kbn-openapi-bundler/src/bundler/ref_resolver/index.ts
@@ -6,11 +6,4 @@
  * Side Public License, v 1.
  */
 
-import { DocumentNode, ResolvedRef } from '../../types';
-import { InlinableRefNode } from '../types';
-
-export function inlineRef(node: DocumentNode, resolvedRef: ResolvedRef): void {
-  Object.assign(node, resolvedRef.refNode);
-
-  delete (node as InlinableRefNode).$ref;
-}
+export * from './ref_resolver';
