+++ x-pack/plugins/security_solution/common/api/detection_engine/prebuilt_rules/install_prebuilt_rules_and_timelines/install_prebuilt_rules_and_timelines_route.schema.yaml
@@ -5,6 +5,9 @@ info:
 paths:
   /api/detection_engine/rules/prepackaged:
     put:
+      x-labels:
+        - ess
+        - serverless
       operationId: InstallPrebuiltRulesAndTimelines
       x-codegen-enabled: true
       summary: Installs all Elastic prebuilt rules and timelines
