+++ x-pack/plugins/security_solution/common/api/detection_engine/rule_management/bulk_actions/bulk_actions_route.schema.yaml
@@ -5,6 +5,9 @@ info:
 paths:
   /api/detection_engine/rules/_bulk_action:
     post:
+      x-labels:
+        - ess
+        - serverless
       operationId: PerformBulkAction
       x-codegen-enabled: true
       summary: Applies a bulk action to multiple rules
