+++ x-pack/plugins/security_solution/common/api/detection_engine/rule_management/bulk_crud/bulk_create_rules/bulk_create_rules_route.schema.yaml
@@ -4,8 +4,8 @@ info:
   version: '2023-10-31'
 paths:
   /api/detection_engine/rules/_bulk_create:
-    x-labels: [ess]
     post:
+      x-labels: [ess]
       operationId: BulkCreateRules
       x-codegen-enabled: true
       deprecated: true
