+++ x-pack/plugins/security_solution/common/api/detection_engine/rule_management/bulk_crud/bulk_update_rules/bulk_update_rules_route.schema.yaml
@@ -4,8 +4,8 @@ info:
   version: '2023-10-31'
 paths:
   /api/detection_engine/rules/_bulk_update:
-    x-labels: [ess]
     put:
+      x-labels: [ess]
       operationId: BulkUpdateRules
       x-codegen-enabled: true
       deprecated: true
