+++ x-pack/plugins/security_solution/common/api/detection_engine/rule_management/crud/delete_rule/delete_rule_route.schema.yaml
@@ -5,6 +5,9 @@ info:
 paths:
   /api/detection_engine/rules:
     delete:
+      x-labels:
+        - ess
+        - serverless
       operationId: DeleteRule
       x-codegen-enabled: true
       description: Deletes a single rule using the `rule_id` or `id` field.
