+++ x-pack/plugins/security_solution/common/api/detection_engine/rule_management/crud/patch_rule/patch_rule_route.schema.yaml
@@ -5,6 +5,9 @@ info:
 paths:
   /api/detection_engine/rules:
     patch:
+      x-labels:
+        - ess
+        - serverless
       operationId: PatchRule
       x-codegen-enabled: true
       description: Patch a single rule
