+++ x-pack/plugins/security_solution/common/api/detection_engine/rule_management/crud/read_rule/read_rule_route.schema.yaml
@@ -5,6 +5,9 @@ info:
 paths:
   /api/detection_engine/rules:
     get:
+      x-labels:
+        - ess
+        - serverless
       operationId: ReadRule
       x-codegen-enabled: true
       description: Read a single rule
