+++ x-pack/plugins/security_solution/common/api/detection_engine/rule_management/crud/update_rule/update_rule_route.schema.yaml
@@ -5,6 +5,9 @@ info:
 paths:
   /api/detection_engine/rules:
     put:
+      x-labels:
+        - ess
+        - serverless
       operationId: UpdateRule
       x-codegen-enabled: true
       description: Update a single rule
