+++ x-pack/plugins/security_solution/common/api/detection_engine/rule_management/export_rules/export_rules_route.schema.yaml
@@ -6,6 +6,9 @@ paths:
   /api/detection_engine/rules/_export:
     summary: Exports rules to an `.ndjson` file
     post:
+      x-labels:
+        - ess
+        - serverless
       operationId: ExportRules
       x-codegen-enabled: true
       summary: Export rules
