+++ x-pack/plugins/security_solution/common/api/detection_engine/rule_management/find_rules/find_rules_route.schema.yaml
@@ -5,6 +5,9 @@ info:
 paths:
   /api/detection_engine/rules/_find:
     get:
+      x-labels:
+        - ess
+        - serverless
       operationId: FindRules
       x-codegen-enabled: true
       description: Finds rules that match the given query.
