+++ x-pack/plugins/security_solution/common/api/detection_engine/users/suggest_user_profiles_route.schema.yaml
@@ -6,6 +6,9 @@ paths:
   /internal/detection_engine/users/_find:
     summary: Suggests user profiles based on provided search term
     post:
+      x-labels:
+        - ess
+        - serverless
       operationId: SuggestUserProfiles
       x-codegen-enabled: true
       description: Suggests user profiles.
